requirements.txt
streamlit
speechrecognition
openai
requests
pydub
pyaudio
python-dotenv